+++ forui/lib/src/widgets/date_field/input/date_input_controller.dart
@@ -97,7 +97,6 @@ class DateInputController extends InputController {
 
   @visibleForTesting
   void updateFromCalendar() {
-    print('from controller: ${controller.value}');
     if (!mutating) {
       rawValue = TextEditingValue(
         text: switch (controller.value) {